Q:	How many elephants can you fit in a VW Bug? A:	Four.  Two in the front, two in the back.
Q:	How can you tell if an elephant is in your refrigerator?
A:	There's a footprint in the mayo.

Q:	How can you tell if two elephants are in your refrigerator?
A:	There's two footprints in the mayo.

Q:	How can you tell if three elephants are in your refrigerator?
A:	The door won't shut.

Q:	How can you tell if four elephants are in your refrigerator?
A:	There's a VW Bug in your driveway.